backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic-settings==2.5.2
pydantic[email]==2.9.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
celery[redis]==5.4.0
redis==5.1.1
langdetect==1.0.9
PyMuPDF==1.24.11
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.3
openai==1.51.2
anthropic==0.34.2
httpx==0.27.2
pgvector==0.3.5
python-multipart==0.0.12
aiofiles==24.1.0
reportlab==4.2.4
Pillow==10.4.0
python-dateutil==2.9.0
email-validator==2.2.0
